pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "evapro"
version = "0.1.0"
description = "遍历lims数据库，把新下的lims任务单添加到anneva监控"
authors = ["Yuan Zan <yfinddream@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/seqyuan/evapro"
repository = "https://github.com/seqyuan/evapro"
documentation = "https://github.com/seqyuan/evapro#readme"
include = [
    { path = "evapro/config/evaconf.yaml", format = "sdist" },
    { path = "evapro/config/*.yaml", format = ["wheel"] }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Utilities",
]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
colorama = "^0.4.6"
PyYAML = "^6.0"
pandas = "^2.0.0"
pymysql = "^1.0.2"
importlib-resources = "^5.12.0"

[tool.poetry.group.docs.dependencies]
sphinx-rtd-theme = "^1.3.0"
myst-parser = "^2.0.0"
sphinx-copybutton = "^0.5.2"
sphinx-design = "^0.5.0"
sphinx-book-theme = "^1.0.1"

[tool.poetry.scripts]
evapro = "evapro.scripts.evaproject:main"

[tool.poetry.group.dev.dependencies]
black = "^23.0"
flake8 = "^6.0"
mypy = "^1.0"
pytest = "^7.0"
pytest-cov = "^4.0"
